Allow service manager override
On some systems, ansible isn't doing an awesome job of properly detecting the appropriate service manager. It is nice to be able to override this in at least two cases: 1) where an alternative service manager has been installed or is desired for whatever reason, and 2) to address an issue in some versions of ansible where sysvinit is returned as service.